+++ lib/providers/ConfigDataProvider.tsx
@@ -1,13 +1,18 @@
-import type { WidgetOptions } from "@lib/types";
+import type { WidgetOptions, WidgetThemeOptions } from "@lib/types";
 import { type ReactNode, useMemo } from "react";
 import { createSafeContext } from "../utils/create-safe-context";
 import { LocaleProvider } from "./LocalesProvider";
 import { useAxiosInstance, useSyncedState } from "@lib/hooks";
-import { PreludeData } from "@lib/utils";
-import useSWR, { SWRResponse } from "swr";
 import { ComponentRegistry } from "./componentRegistry";
 import AgentIcon from "../static/agent-icon.png";
 import { AgentType } from "@lib/types/schemas";
+
+const defaultTheme: WidgetThemeOptions = {
+  headerStyle: "compact",
+  primaryColor: "hsl(211,65%,59%)",
+  hideInfoBar: false,
+  triggerOffset: "20px"
+}
 
 function useNormalizeOptions(data: WidgetOptions) {
   return useMemo(() => {
@@ -15,6 +20,8 @@ function useNormalizeOptions(data: WidgetOptions) {
       messageArrived: "https://cloud.opencopilot.so/sfx/notification3.mp3",
       ...data.soundEffectFiles,
     };
+
+    const theme = Object.assign({}, defaultTheme, data.theme ?? {})
 
     const bot: AgentType = {
       id: "555",
@@ -34,6 +41,7 @@ function useNormalizeOptions(data: WidgetOptions) {
       pathParams: data.pathParams ?? {},
       queryParams: data.queryParams ?? {},
       user: data.user ?? {},
+      theme,
       soundEffectFiles,
       collectUserData: data.collectUserData ?? false,
       defaultSettings: {
@@ -52,7 +60,6 @@ type WidgetSettings = {
 interface ConfigData extends ReturnType<typeof useNormalizeOptions> {
   http: ReturnType<typeof useAxiosInstance>;
   botToken: string;
-  preludeSWR: SWRResponse<PreludeData | undefined, any>;
   componentStore: ComponentRegistry;
   widgetSettings: WidgetSettings | null;
   setSettings: (settings: Partial<WidgetSettings>) => void;
@@ -82,18 +89,11 @@ export function ConfigDataProvider({
   );
 
   const _data = useNormalizeOptions(data);
-  console.log(data, _data)
   const http = useAxiosInstance({
     apiUrl: _data.apiUrl,
     botToken: _data.token,
   });
-
-  const preludeSWR = useSWR([http.options], async () => {
-    const { data } = await http.apis.fetchPreludeData();
-    return data;
-  }, {
-    errorRetryCount: 3,
-  });
+  
 
   const [widgetSettings, _setSettings] = useSyncedState("open_settings", _data.defaultSettings, "local");
 
@@ -104,7 +104,7 @@ export function ConfigDataProvider({
 
   return (
     <ConfigDataSafeProvider
-      value={{ ..._data, http, preludeSWR, componentStore, widgetSettings, setSettings }}
+      value={{ ..._data, http, componentStore, widgetSettings, setSettings }}
     >
       <LocaleProvider>{children}</LocaleProvider>
     </ConfigDataSafeProvider>