+++ src/designs/basic/screens/ChatScreenHeader.tsx
@@ -1,13 +1,12 @@
-import { useChat, useConfigData, useLocale } from '@lib/index';
-import { Dialog, DialogClose, DialogContent, DialogTrigger } from '@ui/dialog';
-import { Switch } from '@ui/switch';
+import { useChat, useLocale } from '@lib/index';
+import { Dialog, DialogTrigger } from '@ui/dialog';
 import { Tooltip, TooltipContent, TooltipTrigger } from '@ui/tooltip';
-import { RotateCcw, SettingsIcon, XIcon } from 'lucide-react';
+import { RotateCcw, SettingsIcon } from 'lucide-react';
+import { ClearSessionDialogContent, SettingsDialogContent } from './DialogContents';
 
 const HeroImage = "https://cloud.opencopilot.so/widget/hero-image.png";
 
 function SettingsDialog() {
-    const { widgetSettings, setSettings } = useConfigData();
     const locale = useLocale();
     return (
         <Dialog>
@@ -16,44 +15,12 @@ function SettingsDialog() {
                     {locale.get("settings")}
                 </TooltipContent>
                 <TooltipTrigger asChild>
-                    <DialogTrigger className="p-1.5 rounded-full bg-accent/60 text-background flex-shrink-0">
-                        <SettingsIcon className="size-5" />
+                    <DialogTrigger className="p-1 rounded-full text-secondary bg-black/50 hover:brightness-110 flex-shrink-0">
+                        <SettingsIcon className="size-4" />
                     </DialogTrigger>
                 </TooltipTrigger>
             </Tooltip>
-            <DialogContent>
-                <div className="p-3 flex items-center justify-between">
-                    <h2 className="text-sm font-semibold" dir="auto">
-                        {locale.get("settings")}
-                    </h2>
-                    <DialogClose className="bg-transparent text-accent p-1 font-semibold">
-                        <XIcon className="size-4" />
-                    </DialogClose>
-                </div>
-
-                <div className="p-3 space-y-2">
-                    <div className="flex items-center justify-between">
-                        <label htmlFor="persist-session::open" dir="auto">
-                            {locale.get("persist-session")}
-                        </label>
-                        <Switch
-                            id="persist-session::open"
-                            checked={widgetSettings?.persistSession}
-                            onCheckedChange={(c) => setSettings({ persistSession: c })}
-                        />
-                    </div>
-                    <div className="flex items-center justify-between">
-                        <label htmlFor="sfx::open" dir="auto">
-                            {locale.get("sound-effects")}
-                        </label>
-                        <Switch
-                            id="sfx::open"
-                            checked={widgetSettings?.useSoundEffects}
-                            onCheckedChange={(c) => setSettings({ useSoundEffects: c })}
-                        />
-                    </div>
-                </div>
-            </DialogContent>
+            <SettingsDialogContent />
         </Dialog>
     );
 }
@@ -66,8 +33,7 @@ function HeroImageDisplay() {
     );
 }
 
-function ConfirmationDialog() {
-    const { clearSession } = useChat();
+function ClearSessionDialog() {
     const locale = useLocale();
     return (
         <Dialog>
@@ -78,36 +44,12 @@ function ConfirmationDialog() {
                             {locale.get("reset-conversation")}
                         </TooltipContent>
                         <TooltipTrigger asChild>
-                            <DialogTrigger className="p-1.5 rounded-full bg-accent/60 text-background flex-shrink-0">
-                                <RotateCcw className="size-5" />
+                            <DialogTrigger className="p-1 rounded-full text-secondary bg-black/50 hover:brightness-110 flex-shrink-0">
+                                <RotateCcw className="size-4" />
                             </DialogTrigger>
                         </TooltipTrigger>
                     </Tooltip>
-                    <DialogContent>
-                        <div className="p-4">
-                            <h2 className="text-sm" dir="auto">
-                                {locale.get("reset-conversation-confirm")}
-                            </h2>
-                        </div>
-                        <div className="p-4 space-x-3 flex items-center justify-end">
-                            <button
-                                onClick={() => {
-                                    clearSession();
-                                    setOpen(false);
-                                }}
-                                className="bg-rose-400 text-white px-2 py-1 rounded-lg text-sm"
-                                dir="auto"
-                            >
-                                {locale.get("yes")}
-                            </button>
-                            <DialogClose
-                                className="bg-transparent text-accent border px-2 py-1 rounded-lg text-sm"
-                                dir="auto"
-                            >
-                                {locale.get("no")}
-                            </DialogClose>
-                        </div>
-                    </DialogContent>
+                    <ClearSessionDialogContent setOpen={setOpen} />
                 </>
             )}
         </Dialog>
@@ -121,18 +63,12 @@ function InfoSection() {
             <h2
                 className="text-lg font-semibold text-background text-center"
                 dir="auto"
-                style={{
-                    textShadow: "0px 2px 8px rgba(0, 0, 0, 0.12)",
-                }}
             >
                 {locale.get("got-any-questions")}
             </h2>
             <span
                 className="text-sm text-white text-center"
                 dir="auto"
-                style={{
-                    textShadow: "0px 2px 8px rgba(0, 0, 0, 0.12)",
-                }}
             >
                 {locale.get("typical-response-time")}
             </span>
@@ -140,31 +76,34 @@ function InfoSection() {
     );
 }
 
-// Refactored Header Components
 function HeaderChatRunning() {
     return (
         <header className="p-3 gap-2 flex flex-col pb-4">
             <div className="w-full flex items-center justify-between">
                 <SettingsDialog />
                 <HeroImageDisplay />
-                <ConfirmationDialog />
+                <ClearSessionDialog />
             </div>
         </header>
     );
 }
 
 function HeaderChatDidNotStart() {
-
     return (
-        <header className="p-3 gap-2 flex flex-col pb-8">
+        <header className="p-3 gap-2 flex flex-col">
             <div className="w-full flex items-center justify-between">
                 <SettingsDialog />
                 <HeroImageDisplay />
-                <ConfirmationDialog />
+                <ClearSessionDialog />
             </div>
             <InfoSection />
         </header>
     );
 }
 
-export { HeaderChatRunning, HeaderChatDidNotStart };
+function CompactHeader() {
+    const { noMessages } = useChat();
+    return noMessages ? <HeaderChatDidNotStart /> : <HeaderChatRunning />
+}
+
+export { CompactHeader };