+++ src/designs/constants.ts
@@ -1,9 +1,30 @@
-export const cssVars = `
-[--opn-primary:211_65%_59%]
-[--opn-foreground:0_0%_0%]
-[--opn-background:0_0%_100%]
-[--opn-secondary:0_0%_96%]
-[--opn-primary-foreground:217_72%_18%]
-[--opn-accent:0_0_22%]
-[--opn-dark:0_0%_0%]
-`;
+import type { CSSProperties } from "react";
+import tc from "tinycolor2";
+
+interface Colors {
+    primary: string;
+}
+
+export const cssVars = (colors: Colors, _: { triggerOffset: string }) => {
+    const primary = tc(colors.primary).toHslString();
+    return {
+        "--opn-primary": primary,
+
+        "--opn-background": "0 0% 100%",
+        "--opn-foreground": "210 40% 98%",
+
+        "--opn-accent": "210 40% 96.1%",
+        "--opn-accent-foreground": "222.2 47.4% 11.2%",
+        
+        "--opn-destructive": "0 84.2% 60.2%",
+        "--opn-destructive-foreground": "210 40% 98%",
+        
+        "--opn-secondary": "0 0% 96%",
+        "--opn-secondary-foreground": "222.2 47.4% 11.2%",
+        
+        "--opn-input": "214.3 31.8% 91.4%",
+        "--opn-border": "214.3 31.8% 91.4%",
+
+        "--opn-trigger-offset": _.triggerOffset,
+    } as CSSProperties;
+};