+++ src/iframed.tsx
@@ -1,21 +1,15 @@
 import * as PopoverPrimitive from "@radix-ui/react-popover";
 import Iframe from "@uiw/react-iframe";
-import { MessageSquareDot, XIcon } from "lucide-react";
 import { useState } from "react";
 import styles from "../lib/index.css?inline";
 import { Widget } from "./designs/basic";
-import { cssVars } from "./designs/constants";
-import { cn } from "./utils";
+import { PopoverTrigger } from "./designs/basic/PopoverTrigger";
 
 export function IframedWidgetPopover() {
   const [isOpen, setIsOpened] = useState(false);
 
-  const handleClick = () => {
-    setIsOpened(!isOpen);
-  };
-
   return (
-    <PopoverPrimitive.Root>
+    <PopoverPrimitive.Root open={isOpen} onOpenChange={setIsOpened}>
       <style>{styles}</style>
       <PopoverPrimitive.Content
         onInteractOutside={(ev) => ev.preventDefault()}
@@ -50,26 +44,7 @@ export function IframedWidgetPopover() {
           <Widget data-chat-widget className="font-inter size-full" />
         </Iframe>
       </PopoverPrimitive.Content>
-
-      <PopoverPrimitive.PopoverTrigger
-        data-chat-widget
-        className={`${cssVars} shadow-lg hover:brightness-105 size-fit bottom-5 right-5 transition-all z-[200] fixed font-inter rounded-full text-white bg-primary duration-300 ease-in-out transform active:scale-90`}
-        onClick={handleClick}
-      >
-        <div
-          className={cn(
-            "p-3.5 transition-transform duration-300 relative ease-in-out",
-            { "transform scale-110": isOpen },
-          )}
-        >
-          {!isOpen ? (
-            <MessageSquareDot className="size-7" />
-          ) : (
-            <XIcon className="size-7" />
-          )}
-          <span className="absolute top-0 right-0 size-3 bg-emerald-600 border-2 border-white rounded-full" />
-        </div>
-      </PopoverPrimitive.PopoverTrigger>
+      <PopoverTrigger isOpen={isOpen} />
     </PopoverPrimitive.Root>
   );
 }